feat: add custom editor
