+++ web-packages/imgui-demo/src/core/editor-window.ts
@@ -48,3 +48,15 @@ export class EditorWindow {
 
   }
 }
+
+export const editorWindowStore: (new () => EditorWindow)[] = [];
+export function editorWindow () {
+  return (target: any, context?: unknown) => {
+    for (const editorWindowClass of editorWindowStore) {
+      if (editorWindowClass === target) {
+        console.warn(`Class ${target} 重复注册`);
+      }
+    }
+    editorWindowStore.push(target);
+  };
+}
