+++ web-packages/imgui-demo/src/custom-editors/post-process-volume-editor.ts
@@ -1,0 +1,8 @@
+import { PostProcessVolume } from '@galacean/effects';
+import { Editor, editor } from '../panels';
+
+export class PostProcessVolumeEditor extends Editor {
+
+  override onInspectorGUI (): void {
+  }
+}
