+++ web-packages/imgui-demo/src/panels/index.ts
@@ -1,6 +1,6 @@
-export * from './inspector';
 export * from './composition';
 export * from './sequencer';
 export * from './project';
-export * from './editor';
+export * from './main-editor';
 export * from './preview';
+export * from './editor';
